+++ BEIMA.Client/src/pages/Devices/DevicePage.js
@@ -1,6 +1,7 @@
 import { useOutletContext, useParams, useNavigate } from 'react-router-dom';
 import { useEffect, useState, useRef } from "react"
 import {ItemCard} from "../../shared/ItemCard/ItemCard"
+import {error, warning, success} from '../../shared/Notifications/Notification.js'
 import styles from './DevicePage.module.css'
 import { Form, Card, Button, FormControl, Image} from "react-bootstrap";
 import { TiDelete } from "react-icons/ti";
@@ -204,9 +205,17 @@ const DevicePage = () => {
       setEditable(false)
     }
 
-    const deleteDeviceCall = (id) => {
-      deleteDevice(id);
-      navigate('/devices')
+    const deleteDeviceCall = async (id) => {
+      let deleteNotif = await warning("Warning: Device Deletion", `Are you sure you want to delete device ${tag}?`);
+      if(deleteNotif.isConfirmed){
+        let deleteResult = await deleteDevice(id);
+        if(deleteResult.status === 200){
+          success("Device Deletion Successful", "Device " + tag + " successfully deleted.");
+          navigate('/devices');
+        } else {
+          error("Unable to Delete Device", "Deletion of Device " + tag + " failed.");
+        }
+      }
     }
 
     const cancel = () => {
